Clarify README requirements
As Dropbear support hasn't been built in yet (upstream bug issues) I am clarifying requirements so people don't run on a Dropbear setup and get their hopes up. I think I am on the verge of something of it will take time to implement, so in the meantime the README has to be this way.