Bulletproofing OpenSSH Setup
I noticed in certain instances OpenSSH setups may not work correctly. This was mainly due to either incorrect permissions (like Dropbear) or the client deciding not to use the key. To fix this, I copied the config file setup from the Dropbear side, and added a ssh session to set permissions; all just like Dropbear. Lastly, I cleaned up the comments and spacing a little to make it readable.